feat(ui/new_chat_action_menu): debug view
